requirements.txt
gdsfactory>=7.0.0
pyyaml>=6.0
numpy>=1.20.0
matplotlib>=3.5.0
pydantic>=2.0.0
